requirements.txt
Django==2.1.2
ijson==2.3
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
